refs #23 clean up
