+++ src/anime/map_anime_sys.rs
@@ -1,20 +1,13 @@
-use bevy::core::Name;
 use bevy::math::Vec3;
-use bevy::prelude::{
-    Commands, Entity, Event, EventReader, EventWriter, Query, Res, Time, Timer, TimerMode,
-    Transform, With,
-};
+use bevy::prelude::{Event, EventReader, EventWriter, Query, Res, Time, Transform, With};
 use bevy_ecs_tilemap::prelude::TileTextureIndex;
 use bevy_ecs_tilemap::tiles::TilePos;
 use tracy_client::span;
 
-use crate::{
-    TILE_ANIMATION_SPEED, TILE_ANIMATION_TEXTURE_END_IDX, TILE_ANIMATION_TEXTURE_START_IDX,
-    TILE_SIZE,
-};
 use crate::anime::anime_component::{AnimationComponent, AnimationTimer};
-use crate::anime::anime_res::TileAnimationResource;
 use crate::kinetic_entity::PlayerEntityTag;
+use crate::map::tiled_components::TileEntityTag;
+use crate::TILE_SIZE;
 
 #[derive(Event)]
 pub struct TileAnimationEvent {
@@ -37,12 +30,15 @@ pub fn animate_overlapped_tiles_event_based(
 pub fn handle_overlap_event(
     time: Res<Time>,
     mut event_reader: EventReader<TileAnimationEvent>,
-    mut tile_query: Query<(
-        &TilePos,
-        &mut AnimationTimer,
-        &AnimationComponent,
-        &mut TileTextureIndex,
-    )>,
+    mut tile_query: Query<
+        (
+            &TilePos,
+            &mut AnimationTimer,
+            &AnimationComponent,
+            &mut TileTextureIndex,
+        ),
+        With<TileEntityTag>,
+    >,
 ) {
     let _span = span!("tile animation_loadtime event read");
