+++ src/anime/mod.rs
@@ -1,4 +1,5 @@
 pub mod anime_component;
 pub mod anime_res;
-pub mod character_anime_sys;
+pub mod environment_anime_sys;
 pub mod map_anime_sys;
+pub mod overlay_anime_sys;
