+++ src/anime/overlay_anime_sys.rs
@@ -14,34 +14,7 @@ use crate::{
 };
 use crate::anime::anime_component::{AnimationComponent, AnimationTimer};
 use crate::anime::anime_res::{AnimationAssets, AnimationResource};
-use crate::kinetic_entity::{EnvironmentEntityTag, PlayerEntityTag};
-
-// TODO: Figure out dynamic animation asset loading? where the overlay animation asset gets loaded during runtime or something?
-pub fn attach_animations_to_player_entities(
-    mut commands: Commands,
-    overlay_animation_data: Res<AnimationResource>,
-    query: Query<Entity, With<PlayerEntityTag>>,
-) {
-    for entity in query.iter() {
-        commands.entity(entity).with_children(|parent| {
-            parent
-                .spawn(SpriteSheetBundle {
-                    texture: overlay_animation_data.animation_image_handle.clone(),
-                    atlas: TextureAtlas {
-                        layout: overlay_animation_data.animation_texture_atlas.clone(),
-                        index: PLAYER_ENTITY_ANIMATION_TEXTURE_START_IDX,
-                    },
-                    transform: Transform::default(), // gets overwritten by the parent??
-                    ..Default::default()
-                })
-                .insert(AnimationTimer(Timer::from_seconds(
-                    overlay_animation_data.animation.speed,
-                    TimerMode::Repeating,
-                )))
-                .insert(Name::new("PlayerEntityOverlayAnimation"));
-        });
-    }
-}
+use crate::kinetic_entity::PlayerEntityTag;
 
 pub fn insert_overlay_animation_resources_into_world(
     mut commands: Commands,
@@ -70,6 +43,33 @@ pub fn insert_overlay_animation_resources_into_world(
     });
 }
 
+// TODO: Figure out dynamic animation asset loading? where the overlay animation asset gets loaded during runtime or something?
+pub fn attach_animations_to_player_entities(
+    mut commands: Commands,
+    overlay_animation_data: Res<AnimationResource>,
+    query: Query<Entity, With<PlayerEntityTag>>,
+) {
+    for entity in query.iter() {
+        commands.entity(entity).with_children(|parent| {
+            parent
+                .spawn(SpriteSheetBundle {
+                    texture: overlay_animation_data.animation_image_handle.clone(),
+                    atlas: TextureAtlas {
+                        layout: overlay_animation_data.animation_texture_atlas.clone(),
+                        index: PLAYER_ENTITY_ANIMATION_TEXTURE_START_IDX,
+                    },
+                    transform: Transform::default(), // gets overwritten by the parent??
+                    ..Default::default()
+                })
+                .insert(AnimationTimer(Timer::from_seconds(
+                    overlay_animation_data.animation.speed,
+                    TimerMode::Repeating,
+                )))
+                .insert(Name::new("PlayerEntityOverlayAnimation"));
+        });
+    }
+}
+
 pub fn animate_overlay_animations(
     time: Res<Time>,
     overlay_animation_data: Res<AnimationResource>,
@@ -83,30 +83,6 @@ pub fn animate_overlay_animations(
             // AnimationComponent doesnt derive copy or clone this next line doesnt work
             let animation = overlay_animation_data.animation;
 
-            let current_index = overlay_atlas.index;
-
-            let next_index = if current_index == animation.end_idx as usize {
-                animation.start_idx as usize
-            } else {
-                current_index + 1
-            };
-
-            overlay_atlas.index = next_index;
-        }
-    }
-}
-
-pub fn animate_env_entity_animations(
-    time: Res<Time>,
-    mut query: Query<
-        (&mut AnimationTimer, &AnimationComponent, &mut TextureAtlas),
-        With<EnvironmentEntityTag>,
-    >,
-) {
-    for (mut timer, animation, mut overlay_atlas) in query.iter_mut() {
-        timer.0.tick(time.delta());
-
-        if timer.0.just_finished() {
             let current_index = overlay_atlas.index;
 
             let next_index = if current_index == animation.end_idx as usize {
