+++ src/map/mod.rs
@@ -1,2 +1,3 @@
+pub mod tiled_components;
 pub mod tiled_res;
 pub mod tiled_sys;
